Moves vendors into src
Since this repository has become Axis vendor only there is no
longer a need to separate it from the normal source code.

This commit moves the two files into src/ and removes the
meson option 'vendors'.
