+++ lib/README.md
@@ -8,16 +8,13 @@ lib
 |   |   └── plugin.c
 |   └── unthreaded-signing
 |       └── plugin.c
-├── src
-|   ├── includes
-|   |   └── public header files
-|   └── source files
-└── vendors
-    └── axis-communications
-        └── source files
+└── src
+    ├── includes
+    |   └── public header files
+    └── source files
 ```
 
-The library is organized in [source code](./src/), [plugins](./plugins/) and [vendors](./vendors/).
+The library is organized in [source code](./src/) and [plugins](./plugins/).
 The source code includes all necessary source files for both signing and validation, and there is no
 conceptual difference in building the library for signing or for validation.
 
@@ -26,9 +23,3 @@ plugin. The interfaces can be found in
 [signed_video_signing_plugin.h](./src/includes/signed_video_signing_plugin.h). The framework comes
 with both a threaded and an unthreaded signing plugin. When building the library with the meson
 structure in this repository, the library includes that plugin.
-
-Vendor specific code and APIs are typically handling extra metadata added to the SEI/OBU Metadata,
-which needs to be interpreted correctly when validating authenticity. With the meson option `vendor`
-the user can select which vendor(s) to include in the build. Typically, when building for signing
-the vendor for that camera is selected, whereas when building for validation all vendors are
-included. By default, all vendors are added.
