+++ lib/meson.build
@@ -1,4 +1,3 @@
 # Process plugins first to get the plugin sources used in src/
 subdir('plugins')
-subdir('vendors')
 subdir('src')
