--- lib/vendors/axis-communications/meson.build
@@ -1,4 +1,0 @@
-vendor_sources = files(
-  'sv_vendor_axis_communications.c',
-  'sv_vendor_axis_communications_internal.h',
-)
