--- lib/vendors/meson.build
@@ -1,7 +1,0 @@
-vendor_sources = []
-vendorinc = include_directories('.')
-
-if build_with_axis
-  message('build with axis-communications')
-  subdir('axis-communications')
-endif
